+++ NEWS.md
@@ -1,3 +1,8 @@
+# BGmisc 1.4.2
+* Added twinIDs for potter and asoiaf pedigrees
+* Added twinID to simulatePedigree function, and extended to include MZ, DZ, and SS twins.
+* Added a few more tests for simulatePedigree
+
 # BGmisc 1.4.1
 * replaced print with message in all functions
 * Exposed several internal functions to the user